Create common parent for Jpa*AttributeEntity
Closes #10071
